backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
transformers>=4.35.2
python-dotenv>=1.0.0
pytest>=7.4.3
torch>=2.6.0
numpy>=1.26.2
